pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clear-med"
version = "0.1.0"
description = "CLIP-based Evaluation and Analysis for Radiology"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "CLEAR Contributors" }
]
licenses = [{ text = "MIT" }]
keywords = ["medical-imaging", "clip", "radiology", "zero-shot"]
dependencies = [
    "torch>=1.12",
    "torchvision>=0.13",
    "numpy>=1.21",
    "pandas>=1.3",
    "pillow>=9.0",
    "tqdm>=4.64",
    "scikit-learn>=1.0",
    "scipy>=1.7",
    "h5py>=3.6",
    "ftfy>=6.1",
    "regex>=2021.8.3"
]

[project.optional-dependencies]
visualization = ["matplotlib>=3.5"]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.package-data]
"clear" = ["*.gz"]
